requirements.txt
easydict==1.9
Flask==2.2.2
numpy==1.24.0
python-dotenv==0.21.0
requests==2.28.1
torch==1.8.1
tqdm==4.64.1
transformers==4.25.1
